staging: Remove fw-api, qca-wifi-host-cmn and qcacld-3.0
Both the reported and actual receive link speed are taking a major hit when using the LA.UM.8.4.r1 branch.

The reason for this is thus far unknown, but it doesn't appear to be the driver itself, nor its config. It is possible that it simply doesn't play nice with our fw and/or Sony blobs.

We will switch back to the LA.UM.7.4.r1 branch, which doesn't show this issue.